perf: lazy import njit functions
